agents/runtime_cost/requirements.txt
# Pinned to exact versions to prevent future breakage.

# Strands Agents with A2A support
strands-agents==1.29.0

# AgentCore SDK and tools
bedrock-agentcore==1.4.3
strands-agents-tools==0.2.22

# A2A SDK for protocol support
a2a-sdk==0.3.24

# HTTP clients
httpx==0.28.1
requests==2.32.5

# AWS SDK
boto3==1.42.63
botocore==1.42.63

# Web framework for A2A server
fastapi==0.135.1
uvicorn==0.41.0

# MCP client
mcp==1.26.0

# Date parsing (required by robust_date_parser.py)
parsedatetime==2.6
python-dateutil==2.9.0.post0